Request validators and business logic validators are separated.
- Разделить валидаторы внешних контрактов и валидаторы бизнес-правил в разные папки в структуре проекта
